Refactored a bit (split by files)
