+++ test/helpers.py
@@ -1,0 +1,33 @@
+from urlparse import urlparse, urljoin
+from selenium.common.exceptions import NoSuchElementException
+
+class Page:
+    BASE_URL = 'https://target.my.com/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+        self.window_handle = driver.current_window_handle
+
+    def open(self):
+        url = urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+        self.wait_for_load()
+    
+    def wait_for_load(self):
+        return
+
+    def check_element_exists(self, xpath):
+        try:
+            self.driver.find_element_by_xpath(xpath)
+        except NoSuchElementException:
+            return False
+        return True
+
+    def close(self):
+        self.driver.close()
+
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
